aider: Fix import statement error in app.js by loading script as a module in index.html.
# Aider chat conversation:

USER: app.js is showing a console error: Cannot use import statement outside a module

ASSISTANT: To resolve the issue with the import statement in `app.js`, we need to ensure that the script is loaded as a module. This can be done by adding the `type="module"` attribute to the script tag in `index.html`. Here are the changes:

```diff
--- index.html
+++ index.html
@@ ... @@
 <body>
     <canvas id="bouquetCanvas"></canvas>
-    <script src="scripts/app.js"></script>
+    <script type="module" src="scripts/app.js"></script>
 </body>
 </html>
```

This change specifies that `app.js` should be treated as an ES6 module, which allows the use of the `import` statement.